Merges Search into SearchOperator, uses Query and AssetSearch in search, executes search asynchronally
