+++ src/search/asset_search.py
@@ -1,15 +1,24 @@
 """Asset Search."""
-from typing import List
+from dataclasses import dataclass
+from typing import Dict, List
 
 from bpy.types import Context
 
-from .search import SearchResult
 from ..asset.asset_type import AssetType
 from ...config import HANA3D_NAME
 
 
+@dataclass
+class SearchResult(object):
+    """Hana3D search result."""
+
+    view_id: str
+    file_name: str
+    download_url: str
+    asset_type: AssetType
+
 class AssetSearch(object):
-    """Hana3D search information by asset type (TODO: merge this class with `Search`)."""
+    """Hana3D search information by asset type."""
 
     def __init__(self, context: Context, asset_type: AssetType):
         """Create a Search object by asset type.
@@ -21,8 +30,7 @@ class AssetSearch(object):
         self.context = context
         self.asset_type = asset_type
 
-    @property  # noqa : WPS110
-    def results(self) -> List[SearchResult]:  # noqa : WPS110
+    def get_results(self) -> List[SearchResult]:
         """Get search results by asset type.
 
         Returns:
@@ -30,8 +38,7 @@ class AssetSearch(object):
         """
         return self.context.window_manager.get(f'{HANA3D_NAME}_{self.asset_type}_search')
 
-    @property
-    def results_orig(self) -> List[SearchResult]:
+    def get_original_results(self) -> List[Dict]:
         """Get original search results by asset type (TODO: refactor this logic).
 
         Returns:
@@ -39,10 +46,10 @@ class AssetSearch(object):
         """
         return self.context.window_manager.get(f'{HANA3D_NAME}_{self.asset_type}_search_orig')
 
-    @results.setter  # noqa : WPS110
-    def results(self, results: List[SearchResult]):  # noqa : WPS110
+    def set_results(self, results: List[SearchResult]):  # noqa : WPS110
+        self.results = results
         self.context.window_manager[f'{HANA3D_NAME}_{self.asset_type}_search'] = results
 
-    @results_orig.setter
-    def results_orig(self, results_orig: List[SearchResult]):
-        self.context.window_manager[f'{HANA3D_NAME}_{self.asset_type}_search_orig'] = results_orig
+    def set_original_results(self, original_results: List[Dict]):
+        self.original_results = original_results
+        self.context.window_manager[f'{HANA3D_NAME}_{self.asset_type}_search_orig'] = original_results
