+++ src/ui/operators/asset_bar.py
@@ -11,7 +11,7 @@ from mathutils import Vector
 from ..callbacks.asset_bar import draw_callback2d, draw_callback3d
 from ..main import UI
 from ...preferences.preferences import Preferences
-from ...search.search import Search
+from ...search import SearchOperator
 from ...upload import upload
 from .... import search, utils
 from ....config import (
@@ -35,8 +35,9 @@ def get_asset_under_mouse(mousex: float, mousey: float) -> int:
     """
     ui_props = getattr(bpy.context.window_manager, HANA3D_UI)
 
-    search_object = Search(bpy.context)
-    search_results = search_object.results
+    search_object = SearchOperator(bpy.context)
+    asset_type = search_object._get_asset_type_from_ui()
+    search_results = search_object.get_results(asset_type).results
     len_search = len(search_results)
     if search_results is not None:
 
@@ -229,8 +230,9 @@ def update_ui_size(area: bpy.types.Area, region: bpy.types.Region) -> None:
     ui.bar_width = region.width - ui.bar_x - ui.bar_end
     ui.wcount = math.floor((ui.bar_width - 2 * ui.drawoffset) / (ui.thumb_size + ui.margin))
 
-    search_object = Search(bpy.context)
-    search_results = search_object.results
+    search_object = SearchOperator(bpy.context)
+    asset_type = search_object._get_asset_type_from_ui()
+    search_results = search_object.get_results(asset_type).results
     if search_results is not None and ui.wcount > 0:
         ui.hcount = min(
             user_preferences.max_assetbar_rows,
@@ -280,10 +282,11 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
 
     def search_more(self):
         """Search more results."""
-        search_object = Search(bpy.context)
-        search_results_orig = search_object.results_orig
+        search_object = SearchOperator(bpy.context)
+        asset_type = search_object._get_asset_type_from_ui()
+        search_results_orig = search_object.get_results_original(asset_type)
         if search_results_orig is not None and search_results_orig.get('next') is not None:
-            search.search(get_next=True)
+            search.async_execute(get_next=True)
 
     def exit_modal(self):
         """Exit modal."""
@@ -401,9 +404,9 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
 
         # TODO add one more condition here to take less performance.
         scene = bpy.context.scene
-        search_object = Search(context)
-        search_results = search_object.results
-        search_results_orig = search_object.results_orig
+        search_object = SearchOperator(context)
+        asset_type = search_object._get_asset_type_from_ui()
+        search_results = search_object.get_results(asset_type).results
         # If there aren't any results, we need no interaction(yet)
         if search_results is None:
             return {'PASS_THROUGH'}
@@ -472,8 +475,9 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                 bpy.context.window.cursor_set('DEFAULT')
                 return {'PASS_THROUGH'}
 
-            search_object = Search(bpy.context)
-            search_results = search_object.results
+            search_object = SearchOperator(bpy.context)
+            asset_type = search_object._get_asset_type_from_ui()
+            search_results = search_object.get_results(asset_type).results
             len_search = len(search_results)
 
             if not ui_props.dragging:  # noqa: WPS504
@@ -708,7 +712,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
         ui_props = getattr(context.window_manager, HANA3D_UI)
 
         if self.do_search:
-            search.search()
+            search.async_execute()
 
         if ui_props.assetbar_on:
             # we don't want to run the assetbar many times,
@@ -732,11 +736,12 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
         ui_props.assetbar_on = True
         ui_props.turn_off = False
 
-        search_object = Search(bpy.context)
-        search_results = search_object.results
+        search_object = SearchOperator(bpy.context)
+        asset_type = search_object._get_asset_type_from_ui()
+        search_results = search_object.get_results(asset_type).results
         if search_results is None:
-            search_object = Search(bpy.context)
-            search_object.results = []  # noqa: WPS110
+            search_object = SearchOperator(bpy.context)
+            search_object.set_results(asset_type, [])
 
         if context.area.type != 'VIEW_3D':
             logging.warning('View3D not found, cannot run operator')
